+++ src/nirfsg/metadata/enums.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# This file is generated from NI-RFSG API metadata version 25.8.0d9999
+# This file is generated from NI-RFSG API metadata version 25.8.0d197
 enums = {
     'AllowOutOfSpecificationUserSettings': {
         'values': [
@@ -958,7 +958,7 @@ enums = {
         ]
     },
     'ResetWithOptionsStepsToOmit': {
-        'enum_class': 'IntFlag',
+        'class': 'IntFlag',
         'values': [
             {
                 'documentation': {
@@ -1066,7 +1066,7 @@ enums = {
         ]
     },
     'SelfCalibrateRangeStepsToOmit': {
-        'enum_class': 'IntFlag',
+        'class': 'IntFlag',
         'values': [
             {
                 'documentation': {